remove old code generators
