+++ ThorSoft.Optics/ThorSoft.Optics.Generator.IntegrationTests/OpticsGeneratorTests.cs
@@ -2,7 +2,6 @@
 {
     public sealed class OpticsGeneratorTests
     {
-        [FocusProperties]
-        private sealed class TestClass { }
+        
     }
 }
