--- ThorSoft.Optics/ThorSoft.Optics.Generator.Tests/LensGeneratorTests.RecordStruct_PrimaryConstructorWithSingleParameter#LensGenerator.TestStruct.g.verified.cs
@@ -1,12 +1,0 @@
-﻿//HintName: LensGenerator.TestStruct.g.cs
-namespace Test.Module;
-
-partial record struct TestStruct
-{
-
-        public static global::ThorSoft.Optics.Lens<TestStruct, int> PropertyLens =>
-            new global::ThorSoft.Optics.Lens<TestStruct, int>(
-                static (instance) => instance.Property,
-                static (value, instance) => instance with { Property = value });
-
-} // TestStruct
