--- ThorSoft.Optics/ThorSoft.Optics.Generator/Generation/CodeGenerationRequest.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.CodeAnalysis;
-using ThorSoft.Optics.Generator.Util;
-
-namespace ThorSoft.Optics.Generator.Generation
-{
-    internal record class CodeGenerationRequest
-    {
-        public EquatableMemory<Diagnostic> Diagnostics { get; init; } = [];
-    }
-}
