--- ThorSoft.Optics/ThorSoft.Optics.Generator/Generation/Focus/NoInterception.cs
@@ -1,16 +1,0 @@
-﻿using Microsoft.CodeAnalysis;
-
-namespace ThorSoft.Optics.Generator.Generation.Focus
-{
-    internal sealed record class NoInterception : CodeGenerationRequest
-    {
-        public NoInterception()
-        {
-        }
-
-        public NoInterception(Diagnostic diagnostic)
-        {
-            Diagnostics = new([diagnostic]);
-        }
-    }
-}
