--- ThorSoft.Optics/ThorSoft.Optics.Generator/Generation/Lenses/LensToGenerate.cs
@@ -1,9 +1,0 @@
-﻿namespace ThorSoft.Optics.Generator.Generation.Lenses
-{
-    internal sealed record class LensToGenerate
-    {
-        public required string Name { get; init; }
-        public required string Visibility { get; init; }
-        public required string Type { get; init; }
-    }
-}
