--- ThorSoft.Optics/ThorSoft.Optics.Generator/Options/CodeGenerationOptions.cs
@@ -1,7 +1,0 @@
-﻿namespace ThorSoft.Optics.Generator.Options
-{
-    internal sealed record class CodeGenerationOptions
-    {
-        public required bool UseStaticLambdas { get; init; }
-    }
-}
