--- ThorSoft.Optics/ThorSoft.Optics.Generator/Options/GlobalOptions.cs
@@ -1,9 +1,0 @@
-﻿using System.Text;
-
-namespace ThorSoft.Optics.Generator.Options
-{
-    internal static class GlobalOptions
-    {
-        internal static Encoding SourceTextEncoding { get; } = Encoding.UTF8;
-    }
-}
